Update new credits page scripting: count of selected credits
