+++ mtp_cashbook/assets-src/__old__/javascripts/main.js
@@ -10,7 +10,6 @@
   require('sign-up').SignUpChoices.init();
   require('sticky-header').StickyHeader.init();
   require('batch-validation').BatchValidation.init();
-  require('selection-count').SelectionCount.init();
   require('confirm-manual').ConfirmManual.init();
   require('navigation').Navigation.init();
 }());
